moved RNN test, updated code hygine added windows specific command as a tab and added a warning if the rsl_rl version is wrong
